% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmyloGram_gui.R
\name{AmyloGram_gui}
\alias{AmyloGram_gui}
\title{AmyloGram Graphical User Interface}
\usage{
AmyloGram_gui()
}
\description{
Launches graphical user interface that predicts presence of amyloids.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

