% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pep424}
\alias{pep424}
\title{pep424 data set}
\format{a list of 424 peptides (class \code{\link[seqinr]{SeqFastaAA}}).}
\source{
Walsh, I., Seno, F., Tosatto, S.C.E., and Trovato, A. (2014).
\emph{PASTA 2.0: an improved server for protein aggregation prediction}.
Nucleic Acids Research gku399.
}
\usage{
pep424
}
\description{
Benchmark dataset for PASTA 2.0. 5 sequences shorter than 6 amino acids
(1\% of the original dataset) were removed.
}
\keyword{datasets}
