% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ag_model}
\alias{predict.ag_model}
\title{Predict amyloids}
\usage{
\method{predict}{ag_model}(object, newdata, ...)
}
\arguments{
\item{object}{\code{ag_model} object.}

\item{newdata}{\code{list} of sequences (for example as given by
\code{\link[seqinr]{read.fasta}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Recognizes amyloids using AmyloGram algorithm.
}
\examples{
data(AmyloGram_model)
data(pep424)
predict(AmyloGram_model, pep424[17])
}
