% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ag_model}
\alias{print.ag_model}
\title{Print AmyloGram object}
\usage{
\method{print}{ag_model}(x, ...)
}
\arguments{
\item{x}{\code{ag_model} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints \code{ag_model} objects.
}
\examples{
data(AmyloGram_model)
print(AmyloGram_model)
}
