% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ag_prediction}
\alias{print.ag_prediction}
\title{Print AmyloGram prediction}
\usage{
\method{print}{ag_prediction}(x, ...)
}
\arguments{
\item{x}{\code{ag_prediction} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints \code{ag_prediction} objects.
}
\examples{
data(AmyloGram_model)
data(pep424)
pred <- predict(AmyloGram_model, pep424[c(4, 10)])
print(pred)
}
