% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spec_sens}
\alias{spec_sens}
\title{Specificity/sensitivity balance}
\format{a data frame with four columns and 99 rows.}
\source{
Walsh, I., Seno, F., Tosatto, S.C.E., and Trovato, A. (2014).
\emph{PASTA 2.0: an improved server for protein aggregation prediction}.
Nucleic Acids Research gku399.
}
\usage{
spec_sens
}
\description{
Sensitivity, specificity and Matthew's Correlation Coefficient
of AmyloGram for different cutoffs computed on \code{pep424} dataset.
}
\keyword{datasets}
