% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmyloGram-package.R
\docType{package}
\name{AmyloGram-package}
\alias{AmyloGram}
\alias{AmyloGram-package}
\title{Prediction of amyloids}
\description{
Amyloids
}
\author{
Michal Burdukiewicz.

Maintainer: Michal Burdukiewicz <michalburdukiewicz@gmail.com>
}
\references{
Burdukiewicz MJ, Sobczyk P, Roediger S, Duda-Madej A,
Mackiewicz P, Kotulska M. (2016) \emph{Prediction of amyloidogenicity
based on the n-gram analysis}. PeerJ Preprints 4:e2390v1
\url{https://doi.org/10.7287/peerj.preprints.2390v1}
}
\keyword{package}

